+++ CHANGELOG.md
@@ -1,5 +1,6 @@
 # ???? (???? ??, ????)
 
+ * Introduced unit tests
  * Used a trait to help ensure that the `fmt_supported` method on `StdPipe` always has the same
    prototype across platforms.
  * Removed a stray redundant platform conditional